chore: move service mapper to service package
