+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/di/MapperProvider.kt
@@ -81,7 +81,7 @@ import com.wire.kalium.logic.data.publicuser.PublicUserMapperImpl
 import com.wire.kalium.logic.data.session.SessionMapper
 import com.wire.kalium.logic.data.session.SessionMapperImpl
 import com.wire.kalium.logic.data.session.SessionRepository
-import com.wire.kalium.logic.data.team.ServiceMapper
+import com.wire.kalium.logic.data.service.ServiceMapper
 import com.wire.kalium.logic.data.team.TeamMapper
 import com.wire.kalium.logic.data.team.TeamMapperImpl
 import com.wire.kalium.logic.data.user.AvailabilityStatusMapper
